Manage the certificate subject
This code manages adds the ability to create a subject for the
certificates that are not stored in the openssl.cnf.  This is done using
the req(1) option for '-subj', and builds the string using the params
provided by the resource.

Also incluced here is some cleanup and better use of variables.

We need to reload the variables for each instance to keep from using
information from a previously generated instance.  This still needs to
be improved to keep catalog searches to a minimum whilst loading the
correct info for the current provider instance.

This fixes a bug where the keypair would get created and signed by a CA
other than the one requested due to the reuse of variables.